+++ Dockerfile
@@ -8,4 +8,5 @@ COPY requirements.txt requirements.txt
 RUN python3 -m pip install -r requirements.txt && \
     rm requirements.txt
 
-COPY ./tests/data/spiky-polygons.gpkg /home/
+COPY ./tests/data/spiky-polygons.gpkg /home/tests
+COPY spiky.py /home/app
