--- spiky/__init__.py
@@ -1,1 +1,0 @@
-"""Package with tools to remove GeoPackage polygon files."""
